Airline Restrictions
Chef has  bags that she wants to take on a flight. They weigh , , and  kgs respectively. She wants to check-in exactly two of these bags and carry the remaining one bag with her.

The airline restrictions says that the total sum of the weights of the bags that are checked-in cannot exceed  kgs and the weight of the bag which is carried cannot exceed  kgs. Find if Chef can take all the three bags on the flight.