World Record
"Usain Bolt makes a new world record completing  m in  seconds".

This news headline has been echoing in Chef's mind since  when he started his career in sprinting. This time in the  Tokyo Olympics he is determined to make a new world record.

There are mainly  important factors that result in the variation in speed during practice and during competition.

The first factor is a difference in track material which results in variation in the speed at competition with respect to practice by a factor of .

The second factor is wind speed which results in variation in the speed at competition with respect to practice by a factor of .

The final factor is the maximum speed achieved during practice. If the maximum speed during practice is  m/s then in competition it will be  m/s.

Given Chef's max speed  during practice and the factors , find whether he will be able to create a new world record, i.e, can he complete  m in less than  seconds?

Note: The final time displayed by the Olympics clock is showing after rounding the original time to  places of decimal and is the only measure of runners performance.